% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datGroceries.R
\docType{data}
\name{datGroceries}
\alias{datGroceries}
\title{World Health Organization TB data}
\format{
\subsection{\code{datGroceries}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.kaggle.com/datasets/rashikrahmanpritom/groceries-dataset-for-market-basket-analysismba?select=Groceries+data.csv}
}
\usage{
datGroceries
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
