% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGraphAR-package.R
\docType{package}
\name{dGraphAR-package}
\alias{dGraphAR}
\alias{dGraphAR-package}
\title{dGraphAR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jamartinezmoreno/dGraphAR}
  \item Report bugs at \url{...}
}

}
\author{
\strong{Maintainer}: Andrés Martínez \email{jandres.martinez@protonmail.com} (\href{https://orcid.org/0000-0002-8729-1124}{ORCID})

}
\keyword{internal}
