% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGtransaction.R
\name{dGtransaction}
\alias{dGtransaction}
\title{Convert a data.frame to transaction format}
\usage{
dGtransaction(datA, transCol, itemCol)
}
\arguments{
\item{datA}{A data.frame which contain a column with the transaction IDs and a column with the items per transaction}

\item{transCol}{An integer denoting the column's index that contains the transactions IDs}

\item{itemCol}{An integer denoting the column's index that contains the items}
}
\value{
An S4 class in transaction format
}
\description{
Convert a data.frame to transaction format
}
\examples{
to be written
}
