% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractClosed.R
\name{extractClosed}
\alias{extractClosed}
\title{Title}
\usage{
extractClosed(rawRulesObj)
}
\arguments{
\item{datTrans}{An object of type transaction}
}
\value{
to be written
}
\description{
Title
}
\examples{
to be written

}
