% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterHochbergHLift.R
\name{filterHochbergHLift}
\alias{filterHochbergHLift}
\title{Title}
\usage{
filterHochbergHLift(hyperlifSignObj, minHLift)
}
\arguments{
\item{datTrans}{An object of type transaction}
}
\value{
to be written
}
\description{
Title
}
\examples{
to be written

}
