% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterRelift.R
\name{filterRelift}
\alias{filterRelift}
\title{Title}
\usage{
filterRelift(filtRulesObj, w)
}
\arguments{
\item{datTrans}{An object of type transaction}
}
\value{
to be written
}
\description{
Title
}
\examples{
to be written

}
