% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqItems.R
\name{freqItems}
\alias{freqItems}
\title{Title}
\usage{
freqItems(datTrans)
}
\arguments{
\item{datTrans}{An object of type transaction}
}
\value{
to be written
}
\description{
Title
}
\examples{
to be written

}
