% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperLiftAndTests.R
\name{hyperLiftAndTests}
\alias{hyperLiftAndTests}
\title{Title}
\usage{
hyperLiftAndTests(closedRulesObj, signifLevel)
}
\arguments{
\item{datTrans}{An object of type transaction}
}
\value{
to be written
}
\description{
Title
}
\examples{
to be written

}
