% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFreqItems.R
\name{plotFreqItems}
\alias{plotFreqItems}
\title{Title}
\usage{
plotFreqItems(datFreq, topN)
}
\arguments{
\item{datTrans}{An object of type transaction}
}
\value{
to be written
}
\description{
Title
}
\examples{
to be written

}
