% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawRules.R
\name{rawRules}
\alias{rawRules}
\title{Title}
\usage{
rawRules(datTrans, minNSupport, confidence, minLen, maxLen)
}
\arguments{
\item{datTrans}{An object of type transaction}
}
\value{
to be written
}
\description{
Title
}
\examples{
to be written

}
