% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGtransaction.R
\name{dGtransaction}
\alias{dGtransaction}
\title{Convert a data.frame to transaction format}
\usage{
dGtransaction(datA, transCol, itemCol)
}
\arguments{
\item{datA}{A data.frame which contain a column with the transaction IDs and a column with the items per transaction}

\item{transCol}{An integer denoting the column's index that contains the transactions IDs}

\item{itemCol}{An integer denoting the column's index that contains the items}
}
\value{
An S4 class in transaction format
}
\description{
Convert a data.frame to transaction format
}
\examples{
dat <- data.table::data.table(rbind(c(1, "milk"), c(1, "bread"), c(2, "onions"), c(3, "onions"), c(3, "beer")))
transDat <- dGtransaction(dat, 1, 2)
transDat

transactions in sparse format with
3 transactions (rows) and
4 items (columns)

}
